Desacoplada/requirements.txt
pydantic==2.9.2
psycopg2-binary==2.9.9
boto3==1.34.34
botocore==1.34.34
